moving package name to conform the linting rules
