--- io.openems.edge.meter.accuenergy.acuvim_II/.settings/org.eclipse.core.resources.prefs
@@ -1,9 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/io/openems/edge/meter/accuenergy/acuvim_ii/AccuenergyAccuvimii.java=UTF-8
-encoding//src/io/openems/edge/meter/accuenergy/acuvim_ii/AccuenergyAccuvimiiImpl.java=UTF-8
-encoding//src/io/openems/edge/meter/accuenergy/acuvim_ii/Config.java=UTF-8
-encoding//test/io/openems/edge/meter/accuenergy/acuvim_ii/MeterAccuenergyAcuvimiiTest.java=UTF-8
-encoding//test/io/openems/edge/meter/accuenergy/acuvim_ii/MyConfig.java=UTF-8
-encoding/<project>=UTF-8
-encoding/bnd.bnd=UTF-8
-encoding/readme.adoc=UTF-8
