+++ io.openems.edge.meter.accuenergy.acuvimii/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,9 @@
+eclipse.preferences.version=1
+encoding//src/io/openems/edge/meter/accuenergy/acuvimii/AccuenergyAccuvimii.java=UTF-8
+encoding//src/io/openems/edge/meter/accuenergy/acuvimii/AccuenergyAccuvimiiImpl.java=UTF-8
+encoding//src/io/openems/edge/meter/accuenergy/acuvimii/Config.java=UTF-8
+encoding//test/io/openems/edge/meter/accuenergy/acuvimii/MeterAccuenergyAcuvimiiTest.java=UTF-8
+encoding//test/io/openems/edge/meter/accuenergy/acuvimii/MyConfig.java=UTF-8
+encoding/<project>=UTF-8
+encoding/bnd.bnd=UTF-8
+encoding/readme.adoc=UTF-8
