+++ io.openems.edge.meter.accuenergy.acuvimii/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: OpenEMS Edge io.openems.edge.meter.accuenergy.acuvim_ii
+Bundle-Name: OpenEMS Edge io.openems.edge.meter.accuenergy.acuvimii
 Bundle-Vendor: OpenEMS Association e.V.
 Bundle-License: https://opensource.org/licenses/EPL-2.0
 Bundle-Version: 1.0.0.${tstamp}
