+++ io.openems.edge.meter.accuenergy.acuvimii/src/io/openems/edge/meter/accuenergy/acuvimii/AccuenergyAccuvimii.java
@@ -1,4 +1,4 @@
-package io.openems.edge.meter.accuenergy.acuvim_ii;
+package io.openems.edge.meter.accuenergy.acuvimii;
 
 import org.osgi.service.event.EventHandler;
